pyproject.toml
[tool.poetry]
name = "serefine-backend"
version = "0.1.0"
description = ""
authors = ["muzammil"]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
# Add your dependencies here, for example:
# requests = "^2.25.1"
# numpy = "^1.20.1"
aiohappyeyeballs = "2.4.0"
aiohttp = "3.10.5"
aioitertools = "0.12.0"
aiosignal = "1.3.1"
aiosqlite = "0.20.0"
alembic = "1.13.2"
annotated-types = "0.7.0"
anthropic = "0.28.1"
anyio = "4.4.0"
arize-phoenix = "4.33.0"
arize-phoenix-evals = "0.15.1"
arize-phoenix-otel = "0.4.1"
asgiref = "3.8.1"
async-timeout = "4.0.3"
asyncpg = "0.29.0"
attrs = "24.2.0"
backoff = "2.2.1"
bcrypt = "4.2.0"
beautifulsoup4 = "4.12.3"
boto3 = "1.35.12"
botocore = "1.35.12"
build = "1.2.1"
cachetools = "5.5.0"
certifi = "2024.8.30"
charset-normalizer = "3.3.2"
chroma-hnswlib = "0.7.6"
chromadb = "0.5.5"
click = "8.1.7"
coloredlogs = "15.0.1"
dataclasses-json = "0.6.7"
deprecated = "1.2.14"
dirtyjson = "1.0.8"
distro = "1.9.0"
exceptiongroup = "1.2.2"
fastapi = "0.112.2"
filelock = "3.15.4"
flatbuffers = "24.3.25"
frozenlist = "1.4.1"
fsspec = "2024.6.1"
google-auth = "2.34.0"
googleapis-common-protos = "1.65.0"
graphql-core = "3.2.3"
greenlet = "3.0.3"
grpcio = "1.66.1"
h11 = "0.14.0"
hdbscan = "0.8.38.post1"
httpcore = "1.0.5"
httptools = "0.6.1"
httpx = "0.27.2"
huggingface-hub = "0.24.6"
humanfriendly = "10.0"
idna = "3.8"
importlib-metadata = "8.4.0"
importlib-resources = "6.4.4"
jinja2 = "3.1.4"
jiter = "0.5.0"
jmespath = "1.0.1"
joblib = "1.4.2"
kubernetes = "30.1.0"
llama-cloud = "0.0.15"
llama-index = "0.11.4"
llama-index-agent-openai = "0.3.0"
llama-index-callbacks-arize-phoenix = "0.2.1"
llama-index-cli = "0.3.0"
llama-index-core = "0.11.4"
llama-index-embeddings-huggingface = "0.3.1"
llama-index-embeddings-openai = "0.2.4"
llama-index-indices-managed-llama-cloud = "0.3.0"
llama-index-legacy = "0.9.48.post3"
llama-index-llms-anthropic = "0.2.1"
llama-index-llms-bedrock = "0.2.1"
llama-index-llms-openai = "0.2.2"
llama-index-multi-modal-llms-openai = "0.2.0"
llama-index-program-openai = "0.2.0"
llama-index-question-gen-openai = "0.2.0"
llama-index-readers-file = "0.2.0"
llama-index-readers-llama-parse = "0.3.0"
llama-index-vector-stores-chroma = "0.2.0"
llama-parse = "0.5.1"
llvmlite = "0.43.0"
mako = "1.3.5"
markdown-it-py = "3.0.0"
markupsafe = "2.1.5"
marshmallow = "3.22.0"
mdurl = "0.1.2"
minijinja = "2.2.0"
mmh3 = "4.1.0"
monotonic = "1.6"
mpmath = "1.3.0"
multidict = "6.0.5"
mypy-extensions = "1.0.0"
nest-asyncio = "1.6.0"
networkx = "3.3"
nltk = "3.9.1"
numba = "0.60.0"
numpy = "1.26.4"
nvidia-cublas-cu12 = "12.1.3.1"
nvidia-cuda-cupti-cu12 = "12.1.105"
nvidia-cuda-nvrtc-cu12 = "12.1.105"
nvidia-cuda-runtime-cu12 = "12.1.105"
nvidia-cudnn-cu12 = "9.1.0.70"
nvidia-cufft-cu12 = "11.0.2.54"
nvidia-curand-cu12 = "10.3.2.106"
nvidia-cusolver-cu12 = "11.4.5.107"
nvidia-cusparse-cu12 = "12.1.0.106"
nvidia-nccl-cu12 = "2.20.5"
nvidia-nvjitlink-cu12 = "12.6.68"
nvidia-nvtx-cu12 = "12.1.105"
oauthlib = "3.2.2"
onnxruntime = "1.19.2"
openai = "1.43.0"
openinference-instrumentation = "0.1.15"
openinference-instrumentation-langchain = "0.1.27"
openinference-instrumentation-llama-index = "3.0.0"
openinference-instrumentation-openai = "0.1.14"
openinference-semantic-conventions = "0.1.10"
opentelemetry-api = "1.27.0"
opentelemetry-exporter-otlp = "1.27.0"
opentelemetry-exporter-otlp-proto-common = "1.27.0"
opentelemetry-exporter-otlp-proto-grpc = "1.27.0"
opentelemetry-exporter-otlp-proto-http = "1.27.0"
opentelemetry-instrumentation = "0.48b0"
opentelemetry-instrumentation-asgi = "0.48b0"
opentelemetry-instrumentation-fastapi = "0.48b0"
opentelemetry-proto = "1.27.0"
opentelemetry-sdk = "1.27.0"
opentelemetry-semantic-conventions = "0.48b0"
opentelemetry-util-http = "0.48b0"
orjson = "3.10.7"
overrides = "7.7.0"
packaging = "24.1"
pandas = "2.2.2"
pillow = "10.4.0"
posthog = "3.6.3"
protobuf = "4.25.4"
psutil = "6.0.0"
psycopg2-binary = "2.9.9"
pyarrow = "17.0.0"
pyasn1 = "0.6.0"
pyasn1-modules = "0.4.0"
pydantic = "2.8.2"
pydantic-core = "2.20.1"
pygments = "2.18.0"
pyjwt = "2.9.0"
pynndescent = "0.5.13"
pypdf = "4.3.1"
pypika = "0.48.9"
pyproject-hooks = "1.1.0"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.0.1"
python-multipart = "0.0.9"
pytz = "2024.1"
pyyaml = "6.0.2"
regex = "2024.7.24"
requests = "2.32.3"
requests-oauthlib = "2.0.0"
rich = "13.8.0"
rsa = "4.9"
s3transfer = "0.10.2"
safetensors = "0.4.4"
scikit-learn = "1.5.1"
scipy = "1.14.1"
sentence-transformers = "3.0.1"
shellingham = "1.5.4"
six = "1.16.0"
sniffio = "1.3.1"
soupsieve = "2.6"
sqlalchemy = "2.0.33"
sqlean-py = "3.45.1"
starlette = "0.38.4"
strawberry-graphql = "0.236.0"
striprtf = "0.0.26"
sympy = "1.13.2"
tenacity = "8.5.0"
threadpoolctl = "3.5.0"
tiktoken = "0.7.0"
tokenizers = "0.19.1"
tomli = "2.0.1"
torch = "2.4.0"
tqdm = "4.66.5"
transformers = "4.44.2"
triton = "3.0.0"
typer = "0.12.5"
typing-inspect = "0.9.0"
typing-extensions = "4.12.2"
tzdata = "2024.1"
umap-learn = "0.5.6"
urllib3 = "2.2.2"
uvicorn = "0.30.6"
uvloop = "0.20.0"
watchfiles = "0.24.0"
websocket-client = "1.8.0"
websockets = "13.0.1"
wrapt = "1.16.0"
yarl = "1.9.8"
zipp = "3.20.1"

[tool.poetry.dev-dependencies]
# Add your development dependencies here, for example:
# pytest = "^6.2.2"

[tool.poetry.group.dev.dependencies]
pylint = "^3.3.0"
astroid = "^3.3.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
